requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
